feat(vcs-plugin): Add "Dummy" VCS
This "Dummy" VCS plugin acts as a fallback, should no supported VCS be detected,
as it is of `VcsType.UNKOWN` and has priority `0`. However, it provides the all
required interfaces of a `VersionControlSystem` for ORT to function as if a VCS
were present. This makes it possible to scan software projects with no version
control or unsupported VCSs.

The plugin is kept simple and provides mostly hard coded return values.
- It's `DummyWorkingTree` is empty and only knows revision `main`.
- As the `RemoteURL` the tree returns the local `workingDir` (given via `-i`).
- Upon calling `initWorkingTree`, the plugin copies content of the `vcs.url`
  into the appropriate cache directory. So ORT can scan the project as is.

Signed-off-by: Jens Keim <jens.keim@forvia.com>