feat(Provenance): Add `DirectoryProvenance` as a `LocalProvenance`
In contrast to the previously added `RemoteProvenance` stands the
`LocalProvenance`, which has no remote source, but instead references
a local input of some kind.
The `DirectoryProvenance` references a local project directory,
which is lacking supported (remote) version control.
It is defined by its canonical path only.

Since `DirectoryProvenance` handling is not yet implemented, the new
class CAN NOT be used yet.

The exception being `hash` and `storageKey`, which both required a
default value, which was set to the `canonicalPath`.
However, since the rest of the code does not handle
`DirectoryProvenance`, this should remain unused for now.

See [1] for more context on the new provenance hierarchy.

[1]: https://github.com/oss-review-toolkit/ort/issues/8803#issuecomment-2236958430

Signed-off-by: Jens Keim <jens.keim@forvia.com>